Making primary data source optional.
