+++ src/main/scala/edu/ie3/simona/config/ConfigFailFast.scala
@@ -8,12 +8,12 @@ package edu.ie3.simona.config
 
 import com.typesafe.config.{Config, ConfigException}
 import com.typesafe.scalalogging.LazyLogging
+import edu.ie3.simona.config.ConfigParams._
 import edu.ie3.simona.config.RuntimeConfig.{
   BaseRuntimeConfig,
   LoadRuntimeConfig,
   StorageRuntimeConfig,
 }
-import edu.ie3.simona.config.ConfigParams._
 import edu.ie3.simona.config.SimonaConfig._
 import edu.ie3.simona.exceptions.InvalidConfigParameterException
 import edu.ie3.simona.io.result.ResultSinkType
@@ -21,7 +21,6 @@ import edu.ie3.simona.model.participant.load.{
   LoadModelBehaviour,
   LoadReferenceType,
 }
-import edu.ie3.simona.service.primary.PrimaryServiceProxy
 import edu.ie3.simona.service.weather.WeatherSource.WeatherScheme
 import edu.ie3.simona.util.CollectionUtils
 import edu.ie3.simona.util.ConfigUtil.CsvConfigUtil.checkBaseCsvParams
@@ -36,6 +35,7 @@ import edu.ie3.util.{StringUtils, TimeUtil}
 import tech.units.indriya.quantity.Quantities
 import tech.units.indriya.unit.Units
 
+import java.text.SimpleDateFormat
 import java.time.ZonedDateTime
 import java.time.format.DateTimeParseException
 import java.util.UUID
@@ -590,9 +590,46 @@ object ConfigFailFast extends LazyLogging {
   }
 
   private def checkPrimaryDataSource(
-      primary: InputConfig.Primary
-  ): Unit =
-    PrimaryServiceProxy.checkConfig(primary)
+      primaryConfig: InputConfig.Primary
+  ): Unit = {
+    val supportedSources =
+      Set("csv", "sql")
+
+    val sourceConfigs = Seq(
+      primaryConfig.couchbaseParams,
+      primaryConfig.csvParams,
+      primaryConfig.influxDb1xParams,
+      primaryConfig.sqlParams,
+    ).filter(_.isDefined).flatten
+
+    if (sourceConfigs.size > 1)
+      throw new InvalidConfigParameterException(
+        s"${sourceConfigs.size} time series source types defined. " +
+          s"Please define only one type!\nAvailable types:\n\t${supportedSources.mkString("\n\t")}"
+      )
+    else if (sourceConfigs.isEmpty) {
+      logger.warn("No primary data source configured.")
+    } else {
+      sourceConfigs.headOption match {
+        case Some(csvParams: TimeStampedCsvParams) =>
+          // note: if inheritance is supported by tscfg,
+          // the following method should be called for all different supported sources!
+          checkTimePattern(csvParams.timePattern)
+        case Some(sqlParams: TimeStampedSqlParams) =>
+          checkTimePattern(sqlParams.timePattern)
+        case Some(x) =>
+          throw new InvalidConfigParameterException(
+            s"Invalid configuration '$x' for a time series source.\nAvailable types:\n\t${supportedSources
+                .mkString("\n\t")}"
+          )
+        case None =>
+          throw new InvalidConfigParameterException(
+            s"No configuration for a time series mapping source provided.\nPlease provide one of the available sources:\n\t${supportedSources
+                .mkString("\n\t")}"
+          )
+      }
+    }
+  }
 
   private def checkWeatherDataSource(
       weatherDataSourceCfg: InputConfig.WeatherDatasource
@@ -976,4 +1013,22 @@ object ConfigFailFast extends LazyLogging {
         )
     }
   }
+
+  /** Check the validity of the given time pattern.
+    * @param dtfPattern
+    *   That should be checked.
+    */
+  private def checkTimePattern(dtfPattern: String): Unit =
+    Try {
+      new SimpleDateFormat(dtfPattern)
+    } match {
+      case Failure(exception) =>
+        throw new InvalidConfigParameterException(
+          s"Invalid timePattern '$dtfPattern' found. Please provide a valid pattern!" +
+            s"\nException: $exception"
+        )
+      case Success(_) =>
+      // this is fine
+    }
+
 }
